Allow for module exclusions
The plugin determines the modules automatically from the project
dependencies in pom.xml.

For the specific case of using javafx libraries (e.g. `javafx-graphics`)
the dependency is a placeholder for the actual dependency
javafx-graphics-win-<version>.jar.
The library designers have decided to not make those bogus dependency
actual modules and they are configured as automatic module. If those get in
`jlink`'s module list, it will exit with an error.

Allow a way to exclude modules such that maven-jlink-plugin can be used
with those projects.