[MJLINK-85] Fix issue with single quoting every argument
Each argument is first double quoted, but because of the default
configuration of `org.apache.maven.shared.utils.cli.shell.Shell`
is also single quoted on top. This leads to a shell command line
which is not executable.

To fix this, leave the double quoting of each argument as it was,
disable the single quoting of each argument with `Shell` configuration
and wrap all the arguments as a whole string with single quotes, so
that they are passed into `bin/sh -c`, example:

```
bin/sh -c '"--module-path" "foo:bar" "--add-modules" "mod1,mod2"'
```